Chapter 2: Vectors
They're just `f64`s for now. I like the idea of making it generic, but
the level of complexity for the operator type constraints is a little
much. It would be the _right_ way, I think, in the long run.

I'll add Points later, once I know what operations I'll need.