Chapter 3: Rays and Points
There's a discrepency between the what the book says and it's sample
code. If our `color`/`colour` function normalizes the input ray's
direction, it doesn't get coloured correctly. The code also makes
things a green rather than just blue -- the `end` of our linear
interpolation in `colour`. I'm guessing the ebook body is out of sync
with the text.